Sat, May 11, 2024, 11:17 PM -03:00

+++ calculator.c
@@ -6,7 +6,7 @@ int main(void)
     long x = get_long("X ");
     long y = get_long("Y ");
 
-    float z = (float) x / (float) y;
+    double z = (double) x / (double) y;
     printf("%.20f\n", z);
 }
